updated signature checking of collections

+++ backend/collection_handler.py
@@ -4,6 +4,10 @@ import hashlib
 import time
 import base64
 import datetime
+from sys import path
+path.append(os.path.expanduser('~/PyBitmessage/src'))
+import addresses
+
 
 
 class CollectionHandler:
@@ -18,8 +22,9 @@ class CollectionHandler:
         :param address:  the address this collection came from
         :return: True if the signatures match, False otherwise
         """
+        to_status, to_address_version_number, to_stream_number, to_ripe = addresses.decodeAddress(address)
+        h = hashlib.sha256(to_ripe.encode('hex') + fj_message['payload']).hexdigest()
 
-        h = hashlib.sha256(fj_message["original_sender"] + fj_message['payload']).hexdigest()
         if h == fj_message["signature"]:
             print "Signature Verified"
             return True
@@ -75,7 +80,7 @@ class CollectionHandler:
 
                 # Trying to filter out non collection messages
                 # TODO Change this filtering technique?
-                if "payload" in json_decode and self._check_signature(json_decode, address):
+                if "payload" in json_decode:
 
                     payload = json_decode["payload"]
                     try:
@@ -83,43 +88,40 @@ class CollectionHandler:
                     except (ValueError, TypeError):
                         print "Contents of FJ Message invalid or corrupted"
                         continue
+                    if self._check_signature(json_decode, payload["address"]):
+                        # Grabbing the text representations of the documents and keywords and rebuilding them
+                        keywords = []
+                        docs = []
+                        for key in payload["keywords"]:
+                            keywords.append(Keyword(name=key[1], id=key[0]))
+                        for doc in payload["documents"]:
+                            docs.append(Document(collection_address=doc[0], description=doc[1], hash=doc[2], title=doc[3]))
 
-                    # Grabbing the text representations of the documents and keywords and rebuilding them
-                    keywords = []
-                    docs = []
-                    for key in payload["keywords"]:
-                        keywords.append(Keyword(name=key[1], id=key[0]))
-                    for doc in payload["documents"]:
-                        docs.append(Document(collection_address=doc[0], description=doc[1], hash=doc[2], title=doc[3]))
+                        collection_model = Collection(
+                            title=payload["title"],
+                            description=payload["description"],
+                            merkle='test',
+                            address=payload["address"],
+                            version=payload["version"],
+                            btc=payload["btc"],
+                            keywords=keywords,
+                            documents=docs,
+                            creation_date=datetime.datetime.strptime(payload["creation_date"], "%A, %d. %B %Y %I:%M%p"),
+                            oldest_date=datetime.datetime.strptime(payload["oldest_date"], "%A, %d. %B %Y %I:%M%p")
+                        )
 
-                    collection_model = Collection(
-                        title=payload["title"],
-                        description=payload["description"],
-                        merkle='test',
-                        address=payload["address"],
-                        version=payload["version"],
-                        btc=payload["btc"],
-                        keywords=keywords,
-                        documents=docs,
-                        creation_date=datetime.datetime.strptime(payload["creation_date"], "%A, %d. %B %Y %I:%M%p"),
-                        oldest_date=datetime.datetime.strptime(payload["oldest_date"], "%A, %d. %B %Y %I:%M%p")
-                    )
-
-                    insert_new_collection(collection_model)
-                    self.connection.delete_message(message['msgid'])
-                    print "Collection cached"
-                    return True
+                        insert_new_collection(collection_model)
+                        self.connection.delete_message(message['msgid'])
+                        print "Collection cached"
+                        return True
 
         print "Could not import collection"
 
     def publish_collection(self, collection, address=''):
         """
         Publishes the given to collection to the bitmessage network
-        :param title: the title of the collection
-        :param description: the description of the collection
-        :param keywords: the keywords that apply to this collection
-        :param documents: the FreeNet document addresses
-        :param address_label: the label for the new Bit Message address
+        :param collection: the collection to be published
+        :param address: the address to send the collection to
         """
 
         collection_payload = self._collection_tojson(collection)
